data: rename collection data to accounts
"accounts" is more specific and better reflects the contents of this
collection. Here I use term "collection" to refer to a set of data of
(mostly) the same type and structure. In our case, it is a file
hierarchy with time series files storing historical metrics for various
social media accounts.