tsu: add entry --replaces switch for alt hosts
It is easier to collect numbers from alternative front-ends to some
platforms because they load faster. For example, nitter.net is much
easier to work with than twitter.com.

Make the existing hard-coded replacement a configurable CLI switch. It
would be more convenient for the user to read it from a config file but
a CLI switch is less code to add so keeping it simple for now.