data: meta: Use update_interval feature
`update_interval` was added in tsu commit 2ba20bb. The value is in
seconds, if the specified amount of time has passed since metric's
latest record, it will be updated. If omitted, it defaults to 86,400
seconds or roughly 1 day.

Values currently used in `accounts.json` are 604,800 (7 days), 1,296,000
(15 days), 2,592,000 (30 days, most common), and 5,184,000 (60 days).