data: meta: Tweak update_interval to handle delays
If `update_interval` is set to 30 days but I update the metric on 3rd
(delay by 2 days), it won't be offered for update on 1st next month.

Reduce (7, 15, 30, 60) to (5, 13, 26, 54) days respectively. This will
"expire" the metrics earlier and trigger their updating on 1st even if
it was delayed in the previous month by a few days.

Not ideal but should be good enough for now.