tsu: Add ability to skip disabled metrics
Skipping accounts was added earlier but now some metrics need to be
skipped while their accounts (and other metrics in them) remain
enabled.

This granularity is required for automating data collection. Some
metrics cannot be collected automatically and the collectors will need
explicit lists of enabled metrics to collect.

Account and metric metadata access is getting ugly. Consumers that just
need to get the metadata value (`entry_path()`, `entry_file()`) have to
also dig through the dicts, deal with missing and default values, and
understand metadata stored in concrete MetricMeta as well as less
structured dicts. This needs to be refactored and simplified in the
future.